+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a custom event handler to the session timeout modal JS to allow applications to trigger keepalive requests. (AUTHLIB-171)
+
 ## [4.0.0] - 2026-03-17
 
 ### Changed