+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.1.1] - 2026-05-06
+
 ### Fixed
 
 - Extract client IP address correctly when `X-Forwarded-For` header contains a list of IP addresses. (AUTHLIB-172)
@@ -480,7 +482,8 @@ _DO NOT USE: This release contains a bug that prevents application startup in so
 - Hide navbar items based on roles (AUTHLIB-25)
 - Prevent creation of table-based users when table-based auth is disabled (AUTHLIB-36)
 
-[unreleased]: https://github.com/OHSU-OCTRI/authentication-lib/compare/v4.1.0...HEAD
+[unreleased]: https://github.com/OHSU-OCTRI/authentication-lib/compare/v4.1.1...HEAD
+[4.1.1]: https://github.com/OHSU-OCTRI/authentication-lib/compare/v4.1.0...v4.1.1
 [4.1.0]: https://github.com/OHSU-OCTRI/authentication-lib/compare/v4.0.0...v4.1.0
 [4.0.0]: https://github.com/OHSU-OCTRI/authentication-lib/compare/v3.2.0...v4.0.0
 [3.2.0]: https://github.com/OHSU-OCTRI/authentication-lib/compare/v3.1.0...v3.2.0